requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.2
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
dotenv==0.9.9
feedparser==6.0.12
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
magic-filter==1.0.12
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
pillow==12.1.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
regex==2026.1.15
requests==2.32.5
setuptools==80.10.1
sgmllib3k==1.0.0
six==1.17.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
yarl==1.22.0
